refactor: 💡 rename bar to header
