+++ src/pages/home/components/HomeComponent.spec.js
@@ -1,0 +1,6 @@
+import { shallowMount } from '@vue/test-utils';
+import HomeComponent from './HomeComponent.vue';
+
+describe('HomeComponent.vue', () => {
+  const wrapper = shallowMount(HomeComponent);
+});
