+++ src/pages/home/containers/HomeContainer.spec.js
@@ -1,0 +1,6 @@
+import { shallowMount } from '@vue/test-utils';
+import HomeContainer from './HomeContainer.vue';
+
+describe('HomeContainer.vue', () => {
+  const wrapper = shallowMount(HomeContainer);
+});
