--- src/pages/navigation-bar/components/index.ts
@@ -1,1 +1,0 @@
-export { default as NavigationBarComponent } from './NavigationBarComponent.vue';
