--- src/pages/navigation-bar/containers/index.ts
@@ -1,1 +1,0 @@
-export { default as NavigationBarPage } from './NavigationBarContainer.vue';
