+++ src/pages/navigation-header/components/index.ts
@@ -1,0 +1,1 @@
+export { default as NavigationHeaderComponent } from './NavigationHeaderComponent.vue';
