+++ src/pages/navigation-header/containers/index.ts
@@ -1,0 +1,1 @@
+export { default as NavigationHeaderPage } from './NavigationHeaderContainer.vue';
